hugging_face/embeddings/embedding-requirements.txt
chromadb
# fastembed-gpu
pillow
fastembed
